backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
pydantic
sqlalchemy
httpx
python-dotenv

